+++ pkg/scm/github/mock_github/mock_client.go
@@ -13,6 +13,7 @@ import (
 	context "context"
 	reflect "reflect"
 
+	scm "github.com/xyzjace/terraplane/pkg/scm"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -70,6 +71,21 @@ func (mr *MockClientMockRecorder) GetFile(ctx, repo, path, revision any) *gomock
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetFile", reflect.TypeOf((*MockClient)(nil).GetFile), ctx, repo, path, revision)
 }
 
+// ListIssueComments mocks base method.
+func (m *MockClient) ListIssueComments(ctx context.Context, repo string, prNumber int) ([]scm.Note, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "ListIssueComments", ctx, repo, prNumber)
+	ret0, _ := ret[0].([]scm.Note)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ListIssueComments indicates an expected call of ListIssueComments.
+func (mr *MockClientMockRecorder) ListIssueComments(ctx, repo, prNumber any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListIssueComments", reflect.TypeOf((*MockClient)(nil).ListIssueComments), ctx, repo, prNumber)
+}
+
 // ReactToComment mocks base method.
 func (m *MockClient) ReactToComment(ctx context.Context, repo string, commentID int, reaction string) error {
 	m.ctrl.T.Helper()
@@ -82,6 +98,20 @@ func (m *MockClient) ReactToComment(ctx context.Context, repo string, commentID 
 func (mr *MockClientMockRecorder) ReactToComment(ctx, repo, commentID, reaction any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ReactToComment", reflect.TypeOf((*MockClient)(nil).ReactToComment), ctx, repo, commentID, reaction)
+}
+
+// UpdateComment mocks base method.
+func (m *MockClient) UpdateComment(ctx context.Context, repo string, commentID int, body string) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "UpdateComment", ctx, repo, commentID, body)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// UpdateComment indicates an expected call of UpdateComment.
+func (mr *MockClientMockRecorder) UpdateComment(ctx, repo, commentID, body any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateComment", reflect.TypeOf((*MockClient)(nil).UpdateComment), ctx, repo, commentID, body)
 }
 
 // WriteComment mocks base method.